Change double's to floats, tested hover mode with this version.

+++ src/math/pi_controller.h
@@ -5,13 +5,13 @@
 class PI_Controller
 {
 public:
-    PI_Controller(double sampling_time,
-                  double gain_proportional,
-                  double time_constant_inverse,
-                  double gain_correction,
-                  double feed_fwd,
-                  double limit_lower,
-                  double limit_upper ):
+    PI_Controller(float sampling_time,
+                  float gain_proportional,
+                  float time_constant_inverse,
+                  float gain_correction,
+                  float feed_fwd,
+                  float limit_lower,
+                  float limit_upper ):
         _sampling_time(sampling_time),
         _gain_proportional(gain_proportional),
         _time_constant_inverse(time_constant_inverse),
@@ -23,22 +23,22 @@ public:
         _anti_windup(0)
     {}
 
-    double Update(double error);
+    float Update(float error);
 
     void Reset();
 
 
 private:
-    double _sampling_time;
-    double _gain_proportional;
-    double _time_constant_inverse;
-    double _gain_correction;
-    double _feed_fwd;
-    double _limit_lower;
-    double _limit_upper;
+    float _sampling_time;
+    float _gain_proportional;
+    float _time_constant_inverse;
+    float _gain_correction;
+    float _feed_fwd;
+    float _limit_lower;
+    float _limit_upper;
 
 
-    double _integral_part;
-    double _anti_windup;
+    float _integral_part;
+    float _anti_windup;
 
 };
